♻️ Split Teams slices into per-use-case folders
Each Teams use case (6 endpoints + the PowerLevelUpdated event
handler) now lives in its own folder with one file per class.
Namespaces follow the folder, e.g. Features.Teams.GetTeam.
