+++ src/WebApi/Features/Teams/AddHeroToTeam/AddHeroToTeamRequest.cs
@@ -1,0 +1,3 @@
+namespace SSW.VerticalSliceArchitecture.Features.Teams.AddHeroToTeam;
+
+public record AddHeroToTeamRequest(Guid TeamId, Guid HeroId);
