+++ src/WebApi/Features/Teams/AddHeroToTeam/AddHeroToTeamRequestValidator.cs
@@ -1,0 +1,13 @@
+namespace SSW.VerticalSliceArchitecture.Features.Teams.AddHeroToTeam;
+
+public class AddHeroToTeamRequestValidator : Validator<AddHeroToTeamRequest>
+{
+    public AddHeroToTeamRequestValidator()
+    {
+        RuleFor(v => v.TeamId)
+            .NotEmpty();
+
+        RuleFor(v => v.HeroId)
+            .NotEmpty();
+    }
+}
