+++ src/WebApi/Features/Teams/CompleteMission/CompleteMissionRequest.cs
@@ -1,0 +1,3 @@
+namespace SSW.VerticalSliceArchitecture.Features.Teams.CompleteMission;
+
+public record CompleteMissionRequest(Guid TeamId);
