+++ src/WebApi/Features/Teams/CompleteMission/CompleteMissionRequestValidator.cs
@@ -1,0 +1,10 @@
+namespace SSW.VerticalSliceArchitecture.Features.Teams.CompleteMission;
+
+public class CompleteMissionRequestValidator : Validator<CompleteMissionRequest>
+{
+    public CompleteMissionRequestValidator()
+    {
+        RuleFor(v => v.TeamId)
+            .NotEmpty();
+    }
+}
