+++ src/WebApi/Features/Teams/CreateTeam/CreateTeamRequest.cs
@@ -1,0 +1,3 @@
+namespace SSW.VerticalSliceArchitecture.Features.Teams.CreateTeam;
+
+public record CreateTeamRequest(string Name);
