+++ src/WebApi/Features/Teams/CreateTeam/CreateTeamRequestValidator.cs
@@ -1,0 +1,10 @@
+namespace SSW.VerticalSliceArchitecture.Features.Teams.CreateTeam;
+
+public class CreateTeamRequestValidator : Validator<CreateTeamRequest>
+{
+    public CreateTeamRequestValidator()
+    {
+        RuleFor(v => v.Name)
+            .NotEmpty();
+    }
+}
