+++ src/WebApi/Features/Teams/ExecuteMission/ExecuteMissionRequest.cs
@@ -1,0 +1,3 @@
+namespace SSW.VerticalSliceArchitecture.Features.Teams.ExecuteMission;
+
+public record ExecuteMissionRequest(Guid TeamId, string Description);
