+++ src/WebApi/Features/Teams/GetAllTeams/GetAllTeamsResponse.cs
@@ -1,0 +1,6 @@
+namespace SSW.VerticalSliceArchitecture.Features.Teams.GetAllTeams;
+
+public record GetAllTeamsResponse(List<GetAllTeamsResponse.TeamDto> Teams)
+{
+    public record TeamDto(Guid Id, string Name);
+}
