+++ src/WebApi/Features/Teams/GetTeam/GetTeamEndpoint.cs
@@ -1,17 +1,8 @@
 using SSW.VerticalSliceArchitecture.Common.Domain.Teams;
 
-namespace SSW.VerticalSliceArchitecture.Features.Teams.Endpoints;
+namespace SSW.VerticalSliceArchitecture.Features.Teams.GetTeam;
 
-public record GetTeamResponse(Guid Id, string Name, IEnumerable<GetTeamResponse.GetTeamHeroDto> Heroes)
-{
-    public record GetTeamHeroDto(Guid Id, string Name, string Alias, int PowerLevel, IEnumerable<GetTeamHeroPowerDto> Powers);
-
-    public record GetTeamHeroPowerDto(string Name, int PowerLevel);
-}
-
-public record GetTeamRequest(Guid TeamId);
-
-public class GetTeamEndpoint(ApplicationDbContext dbContext) 
+public class GetTeamEndpoint(ApplicationDbContext dbContext)
     : Endpoint<GetTeamRequest, GetTeamResponse>
 {
     public override void Configure()
@@ -34,10 +25,10 @@ public class GetTeamEndpoint(ApplicationDbContext dbContext)
                 t.Name,
                 t.Heroes.Select(
                     h => new GetTeamResponse.GetTeamHeroDto(
-                        h.Id.Value, 
-                        h.Name, 
-                        h.Alias, 
-                        h.PowerLevel, 
+                        h.Id.Value,
+                        h.Name,
+                        h.Alias,
+                        h.PowerLevel,
                         h.Powers.Select(p => new GetTeamResponse.GetTeamHeroPowerDto(p.Name, p.PowerLevel))
                     )
                 )))
@@ -50,23 +41,5 @@ public class GetTeamEndpoint(ApplicationDbContext dbContext)
         }
 
         await Send.OkAsync(team, ct);
-    }
-}
-
-public class GetTeamRequestValidator : Validator<GetTeamRequest>
-{
-    public GetTeamRequestValidator()
-    {
-        RuleFor(v => v.TeamId)
-            .NotEmpty();
-    }
-}
-
-public class GetTeamSummary : Summary<GetTeamEndpoint>
-{
-    public GetTeamSummary()
-    {
-        Summary = "Get a specific team";
-        Description = "Retrieves detailed information about a team, including all its heroes and their powers.";
     }
 }
