+++ src/WebApi/Features/Teams/GetTeam/GetTeamRequest.cs
@@ -1,0 +1,3 @@
+namespace SSW.VerticalSliceArchitecture.Features.Teams.GetTeam;
+
+public record GetTeamRequest(Guid TeamId);
