+++ src/WebApi/Features/Teams/GetTeam/GetTeamRequestValidator.cs
@@ -1,0 +1,10 @@
+namespace SSW.VerticalSliceArchitecture.Features.Teams.GetTeam;
+
+public class GetTeamRequestValidator : Validator<GetTeamRequest>
+{
+    public GetTeamRequestValidator()
+    {
+        RuleFor(v => v.TeamId)
+            .NotEmpty();
+    }
+}
